debug daa type of buffer len

+++ lgt8f/libraries/E2PROM/EEPROM.h
@@ -214,14 +214,14 @@
 			uint8_t read( uint16_t address, bool real_address_mode = false ) { return lgt_eeprom_read_byte( address, real_address_mode ); }
 			void   write( uint16_t address, uint8_t value, bool real_address_mode = false ) { lgt_eeprom_write_byte( address, value, real_address_mode ); }
 
-			void  read_block( uint8_t *pbuf, uint16_t address, uint8_t len, bool real_address_mode = false ) { lgt_eeprom_read_block( pbuf, address, len, real_address_mode ); }
-			void write_block( uint8_t *pbuf, uint16_t address, uint8_t len, bool real_address_mode = false ) { lgt_eeprom_write_block( pbuf, address, len, real_address_mode ); }
+			void  read_block( uint8_t *pbuf, uint16_t address, uint16_t len, bool real_address_mode = false ) { lgt_eeprom_read_block( pbuf, address, len, real_address_mode ); }
+			void write_block( uint8_t *pbuf, uint16_t address, uint16_t len, bool real_address_mode = false ) { lgt_eeprom_write_block( pbuf, address, len, real_address_mode ); }
 
 			uint32_t read32( uint16_t address ) { return lgt_eeprom_read32( address ); }
 			void    write32( uint16_t address, uint32_t data ) { lgt_eeprom_write32( address, data ); }
 
-			void writeSWM( uint16_t address, uint32_t *pdata, uint8_t len) { lgt_eeprom_writeSWM( address, pdata, len ); }
-			void  readSWM( uint16_t address, uint32_t *pdata, uint8_t len) { lgt_eeprom_readSWM( address, pdata, len ); }
+			void writeSWM( uint16_t address, uint32_t *pdata, uint16_t len) { lgt_eeprom_writeSWM( address, pdata, len ); }
+			void  readSWM( uint16_t address, uint32_t *pdata, uint16_t len) { lgt_eeprom_readSWM( address, pdata, len ); }
 
 			int size( bool theoretical = false ) { return lgt_eeprom_size( theoretical ); }
 			
@@ -240,14 +240,14 @@
 			uint8_t read( uint16_t address ) { return lgt_eeprom_read_byte( address ); }
 			void   write( uint16_t address, uint8_t value ) { lgt_eeprom_write_byte( address, value ); }
 
-			void  read_block( uint8_t *pbuf, uint16_t address, uint8_t len ) { lgt_eeprom_read_block( pbuf, address, len ); }
-			void write_block( uint8_t *pbuf, uint16_t address, uint8_t len ) { lgt_eeprom_write_block( pbuf, address, len ); }
+			void  read_block( uint8_t *pbuf, uint16_t address, uint16_t len ) { lgt_eeprom_read_block( pbuf, address, len ); }
+			void write_block( uint8_t *pbuf, uint16_t address, uint16_t len ) { lgt_eeprom_write_block( pbuf, address, len ); }
 
 			uint32_t read32( uint16_t address ) { return lgt_eeprom_read32( address ); }
 			void    write32( uint16_t address, uint32_t data ) { lgt_eeprom_write32( address, data ); }
 
-			void writeSWM( uint16_t address, uint32_t *pdata, uint8_t len) { lgt_eeprom_writeSWM( address, pdata, len ); }
-			void  readSWM( uint16_t address, uint32_t *pdata, uint8_t len) { lgt_eeprom_readSWM( address, pdata, len ); }
+			void writeSWM( uint16_t address, uint32_t *pdata, uint16_t len) { lgt_eeprom_writeSWM( address, pdata, len ); }
+			void  readSWM( uint16_t address, uint32_t *pdata, uint16_t len) { lgt_eeprom_readSWM( address, pdata, len ); }
 			
 
 			int size( bool theoretical = false ) { return lgt_eeprom_size( theoretical ); }
